# app/static/swagger.json
{
    "info": {
        "description": "Flask Api",
        "version": "1.0.0",
        "title": "Flask Api",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "swagger": "2.0",
    "paths": {
        "/api/v1/tasks": {
            "get": {
                "tags": [
                    "Tasks Request"
                ],
                "summary": "Returns taskRequests",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/taskRequests"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tasks Request"
                ],
                "summary": "Create a new task request system",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/taskRequestPostBody"
                            }
                        }
                    }
                },
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/taskRequestPostBody"
                        }
                    },
                    "400": {
                        "description": "Failed. Bad post data."
                    }
                }
            }
        },
        "/api/v1/tasks/{id}": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "ID of the cat that we want to match",
                    "type": "string"
                }
            ],
            "get": {
                "tags": [
                    "Tasks Request"
                ],
                "summary": "Get task request with given ID",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "Tasks Request id",
                        "schema": {
                            "$ref": "#/components/schemas/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/taskRequest"
                        }
                    },
                    "400": {
                        "description": "Failed. Misunderstood Request."
                    },
                    "404": {
                        "description": "Failed. Task request not found."
                    }
                }
            },
            "put": {
                "summary": "edit a task request by ID",
                "tags": [
                    "Tasks Request"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "Tasks Request id",
                        "schema": {
                            "$ref": "#/components/schemas/id"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/taskRequest"
                        }
                    },
                    "400": {
                        "description": "Failed. Bad post data."
                    }
                }
            },
            "delete": {
                "summary": "Delete Tasks Request by ID",
                "tags": [
                    "Tasks Request"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "Tasks Request Id",
                        "schema": {
                            "$ref": "#/components/schemas/id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/id"
                        }
                    },
                    "400": {
                        "description": "Failed. Misunderstood Request."
                    },
                    "404": {
                        "description": "Failed. Tasks Request not found."
                    }
                }
            }
        }
    },
    "components":{
        "schemas": {
            "id": {
                "properties": {
                    "uuid": {
                        "type": "string"
                    }
                }
            },
            "taskRequestPostBody": {
                "type": "object",
                "properties": {
                     "title": {
                        "type": "string",
                        "format": "string"
                    },
                    "description": {
                        "type": "string",
                        "format": "string"
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "due_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "priority": {
                        "type": "string",
                        "format": "priority"
                    }
                }
            },
            "taskRequest": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number",
                        "format": "number"
                    },
                    "title": {
                        "type": "string",
                        "format": "string"
                    },
                    "description": {
                        "type": "string",
                        "format": "string"
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "due_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "priority": {
                        "type": "string",
                        "format": "priority"
                    }
                }
            },
            "taskRequests": {
            "type": "object",
            "properties": {
                "taskRequest": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/components/schemas/taskRequest"
                    }
                }
            }
        }
    }
  }
}